Serialize and Deserialize Quantities as Objects
It should be possible to serialize an JSON Object with the properties "value" and "unit" into an Quantity.
Furthermore it should be possible to serialize a Quantity into an object with the properties "value" and "unit".
The "unit" property should contain the units symbol.
The "value" property should contain the value of the quantity.