Provide Extension Methods for the Time Quantity to convert it to a duration and the other way around.
The implementation should be able to convert a java duration to a time quantity. The implementation should be able to convert a time quantity to a java duration.No loss in information is allowed for the conversion. The converted duration has to be accurate for a resolution of one nano second.