Upgrade Gradle to the latest version 8.9
The newest version of Gradle broke parts of the current build. Some
deprecated features were also still in use. This PR tries to fix some
of those.

Importing and applying scripts directly has been discouraged for a
while. The same is true for using a `subprojects` block to share
configuration. The recommended way to do this are convention plugins.

It is straightforward for the most part to convert the existing
scripts to convention plugins. The contents of the `subprojects` block
can also be applied in a convention plugin.

To make fixing the remaining issues easier, and help in future
development, the convention plugins are then converted to Kotlin. The
code completion and better typing help find and fix issues.

Some small simplifications have been made, where a direct API is
available instead of workarounds. A bigger change of this kind is
changing dependencies from `implementation` to `api` directly instead
of rewriting their scope later directly in the Maven POM.